requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.3.0
backports.weakref==1.0.post1
base58==2.1.1
bitarray==3.7.1
certifi==2025.8.3
charset-normalizer==3.4.3
ckzg==2.1.4
cytoolz==1.0.1
dacite==1.9.2
dataclasses-json==0.6.7
dotenv==0.9.9
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
frozenlist==1.7.0
grvt-pysdk==0.2.1
hexbytes==1.3.1
hyperliquid-python-sdk==0.19.0
idna==3.10
inflection==0.5.1
lighter-sdk @ git+https://github.com/elliottech/lighter-python.git@7b7c2968fef29c93a21706f9d88f7e9f7232f920
loguru==0.7.3
marshmallow==3.26.1
msgpack==1.1.1
multidict==6.6.4
mypy_extensions==1.1.0
packaging==25.0
parsimonious==0.10.0
propcache==0.3.2
pycryptodome==3.23.0
pydantic==2.11.9
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
regex==2025.9.18
requests==2.32.5
rlp==4.1.0
six==1.17.0
toolz==1.0.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.0.7
websocket-client==1.8.0
websockets==13.1
yarl==1.20.1
